requirements.txt
requests==2.20.1
github3.py==1.2.0
slackclient==1.3.0
flask==1.0.2
pyyaml==3.13
gunicorn==19.9.0
redis==3.2.1